CapstoneProject/requirements.txt
<<<<<<< HEAD
# FastAPI Backend
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Document Processing
pypdf2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
unstructured==0.12.4
python-magic==0.4.27

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# LLM Integration
langchain==0.1.9
langchain-community==0.0.20
openai==1.12.0
ollama==0.1.6
groq==0.4.2

# Automation
celery==5.3.6
redis==5.0.1
apscheduler==3.10.4

# UI
streamlit==1.31.1
plotly==5.18.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3
ruff==0.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
=======
# ── Web Framework ──────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Data Validation ────────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Database ───────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── Vector Store ───────────────────────────────────────────────
chromadb>=0.5.0

# ── Embeddings ─────────────────────────────────────────────────
sentence-transformers>=2.7.0

# ── LLM / RAG ──────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.2.0

# ── Document Parsing ───────────────────────────────────────────
PyPDF2>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# ── User Interface ─────────────────────────────────────────────
streamlit>=1.35.0
requests>=2.31.0

# ── Utilities ──────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=23.0.0

# ── Development & Testing ──────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.4.0
>>>>>>> 99af0b4 (Day 1: Project scaffolding and system design)
